Moved extension on Date type, to its own file in extensions. Formatted and cleaned up code in mainTabBar, and JobHistoryTab files
