+++ Capstone/Models/MainTabs/FirstTimeUserExperience/FirstTimeUserExperienceViewController.swift
@@ -9,13 +9,13 @@
 import UIKit
 
 class FirstTimeUserExperienceViewController: UIViewController {
-    
+    //MARK:- IBOutlets
     @IBOutlet weak var pageControl: UIPageControl!
     @IBOutlet weak var collectionView: UICollectionView!
     
-//    private var screenshots = ["screenshot1", "screenshot2", "screenshot3", "screenshot4"]
+    //    private var screenshots = ["screenshot1", "screenshot2", "screenshot3", "screenshot4"]
     private var screenshots = ["CareerViewlogo"]
-
+    //MARK:- ViewLifeCycles
     override func viewDidLoad() {
         super.viewDidLoad()
         configureView()
@@ -29,14 +29,13 @@ class FirstTimeUserExperienceViewController: UIViewController {
         navigationController?.setNavigationBarHidden(false, animated: animated)
         tabBarController?.tabBar.isHidden = false
     }
+    //MARK:- Private funcs
     private func configureView() {
         self.collectionView.delegate = self
         self.collectionView.dataSource = self
         self.collectionView.register(UINib(nibName: "ScreenshotCollectionViewCellXib", bundle: nil), forCellWithReuseIdentifier: "screenshotCell")
-        
         self.pageControl.currentPage = 0
         self.pageControl.numberOfPages = screenshots.count
-        
     }
     private func updateUserFirstTimeLogin() {
         DatabaseService.shared.updateUserFirstTimeLogin(firstTimeLogin: false) { (result) in
@@ -48,17 +47,16 @@ class FirstTimeUserExperienceViewController: UIViewController {
             }
         }
     }
-
+    //MARK:- IBAction funs
     @IBAction func skipForNowButtonPressed(_ sender: UIButton) {
         UIViewController.showMainAppView()
     }
 }
-
+//MARK:- Extensions
 extension FirstTimeUserExperienceViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         let maxWidth: CGFloat = collectionView.frame.width
         let maxHeight: CGFloat = collectionView.frame.height
-
         return CGSize(width: maxWidth, height: maxHeight)
     }
 }
@@ -70,10 +68,6 @@ extension FirstTimeUserExperienceViewController: UIScrollViewDelegate {
         pageControl.currentPage = Int(scrollView.contentOffset.x) / Int(scrollView.frame.width)
     }
 }
-extension FirstTimeUserExperienceViewController: UICollectionViewDelegate {
-    
-}
-
 extension FirstTimeUserExperienceViewController: UICollectionViewDataSource {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
         return screenshots.count
@@ -81,7 +75,6 @@ extension FirstTimeUserExperienceViewController: UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return 1
     }
-    
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath:
         IndexPath) -> UICollectionViewCell {
         guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "screenshotCell", for: indexPath) as? ScreenshotCollectionViewCell else {
@@ -93,33 +86,33 @@ extension FirstTimeUserExperienceViewController: UICollectionViewDataSource {
             let currentScreenshot = screenshots[0]
             cell.configureCell(screenshot: currentScreenshot)
             cell.infoLabel.text = """
-                                Welcome to CallBack!
-                                Browse your Job History
-                                """
+            Welcome to CallBack!
+            Browse your Job History
+            """
             return cell
         case 1:
             let currentScreenshot = screenshots[1]
             cell.configureCell(screenshot: currentScreenshot)
             cell.infoLabel.text = """
-                                Welcome to CallBack!
-                                Practice Interview Questions
-                                """
+            Welcome to CallBack!
+            Practice Interview Questions
+            """
             return cell
         case 2:
             let currentScreenshot = screenshots[2]
             cell.configureCell(screenshot: currentScreenshot)
             cell.infoLabel.text = """
-                                Welcome to CallBack!
-                                Enter your "S.T.A.R" Stories
-                                """
+            Welcome to CallBack!
+            Enter your "S.T.A.R" Stories
+            """
             return cell
         case 3:
             let currentScreenshot = screenshots[3]
             cell.configureCell(screenshot: currentScreenshot)
             cell.infoLabel.text = """
-                                Welcome to CallBack!
-                                Keep Track of your Job Applications
-                                """
+            Welcome to CallBack!
+            Keep Track of your Job Applications
+            """
             return cell
         default:
             let currentScreenshot = screenshots[0]
@@ -127,5 +120,4 @@ extension FirstTimeUserExperienceViewController: UICollectionViewDataSource {
             return cell
         }
     }
-
 }
